requirements.txt
devpi-client>=3.0.0
flake8
pytest-runner
pytest-cov
pytest-mock
readme-renderer
pathlib2; python_version < "3"
